requirements.txt
fastapi
uvicorn[standart]
motor
pydantic[email]
pymongo
python-dotenv